Introduce support for subplots
This commit introduces support for plotting multiple subplots in the
same canvas. A new function aptly called `subplots` lets the user
specify a grid layout in which to split the main canvas and returns an
object through which the canvas and its subcomponents can be managed for
plotting. Various configuration options are available, such as
specifying extra space at the top or bottom of the canvas for extra pads
which could host e.g. a shared common legend for all the subplots. Two
new classes are also introduced. CMSCanvasManager is the type returned
by the `subplots` function. It stores all the pads created and gives
access to them. Each pad is of type CMSPad, the second main class
introduced, which exposes a plot method to plot graphics onto that
specific pad, without needing to interact with internal ROOT mechanisms
such as gPad.

Fixes https://github.com/cms-cat/cmsstyle/issues/52

Co-authored-by: Tommaso Tedeschi <tedesk95@gmail.com>